Modify the Invoke behavior to look up via metatables, allowing inspection to be defined by users
